RH-SYS/RH-Manager/facial_recognition_system/requirements.txt
opencv-python
face-recognition
numpy
requests
pillow
mysql-connector-python
